Move code to prepare the message view menu to `MessageViewFragment`
