+++ app/ui/legacy/src/main/java/com/fsck/k9/activity/MessageList.kt
@@ -1011,126 +1011,6 @@ open class MessageList :
         return true
     }
 
-    override fun onPrepareOptionsMenu(menu: Menu): Boolean {
-        super.onPrepareOptionsMenu(menu)
-        configureMenu(menu)
-        return true
-    }
-
-    /**
-     * Hide menu items not appropriate for the current context.
-     *
-     * **Note:**
-     * Please adjust the comments in `res/menu/message_list_option.xml` if you change the  visibility of a menu item
-     * in this method.
-     */
-    private fun configureMenu(menu: Menu?) {
-        if (menu == null) return
-
-        // Set visibility of menu items related to the message view
-        if (displayMode == DisplayMode.MESSAGE_LIST || messageViewFragment == null ||
-            !messageViewFragment!!.isInitialized
-        ) {
-            menu.findItem(R.id.next_message).isVisible = false
-            menu.findItem(R.id.previous_message).isVisible = false
-            menu.findItem(R.id.single_message_options).isVisible = false
-            menu.findItem(R.id.delete).isVisible = false
-            menu.findItem(R.id.compose).isVisible = false
-            menu.findItem(R.id.archive).isVisible = false
-            menu.findItem(R.id.move).isVisible = false
-            menu.findItem(R.id.copy).isVisible = false
-            menu.findItem(R.id.spam).isVisible = false
-            menu.findItem(R.id.refile).isVisible = false
-            menu.findItem(R.id.toggle_unread).isVisible = false
-            menu.findItem(R.id.toggle_message_view_theme).isVisible = false
-            menu.findItem(R.id.unsubscribe).isVisible = false
-            menu.findItem(R.id.show_headers).isVisible = false
-        } else {
-            // hide prev/next buttons in split mode
-            if (displayMode != DisplayMode.MESSAGE_VIEW) {
-                menu.findItem(R.id.next_message).isVisible = false
-                menu.findItem(R.id.previous_message).isVisible = false
-            } else {
-                val ref = messageViewFragment!!.messageReference
-                val initialized = messageListFragment != null &&
-                    messageListFragment!!.isLoadFinished
-                val canDoPrev = initialized && !messageListFragment!!.isFirst(ref)
-                val canDoNext = initialized && !messageListFragment!!.isLast(ref)
-                val prev = menu.findItem(R.id.previous_message)
-                prev.isEnabled = canDoPrev
-                prev.icon.alpha = if (canDoPrev) 255 else 127
-                val next = menu.findItem(R.id.next_message)
-                next.isEnabled = canDoNext
-                next.icon.alpha = if (canDoNext) 255 else 127
-            }
-
-            val toggleTheme = menu.findItem(R.id.toggle_message_view_theme)
-            if (generalSettingsManager.getSettings().fixedMessageViewTheme) {
-                toggleTheme.isVisible = false
-            } else {
-                // Set title of menu item to switch to dark/light theme
-                if (themeManager.messageViewTheme === Theme.DARK) {
-                    toggleTheme.setTitle(R.string.message_view_theme_action_light)
-                } else {
-                    toggleTheme.setTitle(R.string.message_view_theme_action_dark)
-                }
-                toggleTheme.isVisible = true
-            }
-
-            if (messageViewFragment!!.isOutbox) {
-                menu.findItem(R.id.toggle_unread).isVisible = false
-            } else {
-                // Set title of menu item to toggle the read state of the currently displayed message
-                val drawableAttr = if (messageViewFragment!!.isMessageRead) {
-                    menu.findItem(R.id.toggle_unread).setTitle(R.string.mark_as_unread_action)
-                    intArrayOf(R.attr.iconActionMarkAsUnread)
-                } else {
-                    menu.findItem(R.id.toggle_unread).setTitle(R.string.mark_as_read_action)
-                    intArrayOf(R.attr.iconActionMarkAsRead)
-                }
-                val typedArray = obtainStyledAttributes(drawableAttr)
-                menu.findItem(R.id.toggle_unread).icon = typedArray.getDrawable(0)
-                typedArray.recycle()
-            }
-
-            menu.findItem(R.id.delete).isVisible = K9.isMessageViewDeleteActionVisible
-
-            // Set visibility of copy, move, archive, spam in action bar and refile submenu
-            if (messageViewFragment!!.isCopyCapable) {
-                menu.findItem(R.id.copy).isVisible = K9.isMessageViewCopyActionVisible
-                menu.findItem(R.id.refile_copy).isVisible = true
-            } else {
-                menu.findItem(R.id.copy).isVisible = false
-                menu.findItem(R.id.refile_copy).isVisible = false
-            }
-
-            if (messageViewFragment!!.isMoveCapable) {
-                val canMessageBeArchived = messageViewFragment!!.canMessageBeArchived()
-                val canMessageBeMovedToSpam = messageViewFragment!!.canMessageBeMovedToSpam()
-
-                menu.findItem(R.id.move).isVisible = K9.isMessageViewMoveActionVisible
-                menu.findItem(R.id.archive).isVisible = canMessageBeArchived && K9.isMessageViewArchiveActionVisible
-                menu.findItem(R.id.spam).isVisible = canMessageBeMovedToSpam && K9.isMessageViewSpamActionVisible
-
-                menu.findItem(R.id.refile_move).isVisible = true
-                menu.findItem(R.id.refile_archive).isVisible = canMessageBeArchived
-                menu.findItem(R.id.refile_spam).isVisible = canMessageBeMovedToSpam
-            } else {
-                menu.findItem(R.id.move).isVisible = false
-                menu.findItem(R.id.archive).isVisible = false
-                menu.findItem(R.id.spam).isVisible = false
-
-                menu.findItem(R.id.refile).isVisible = false
-            }
-
-            if (messageViewFragment!!.isOutbox) {
-                menu.findItem(R.id.move_to_drafts).isVisible = true
-            }
-
-            menu.findItem(R.id.unsubscribe).isVisible = messageViewFragment!!.canMessageBeUnsubscribed()
-        }
-    }
-
     fun setActionBarTitle(title: String, subtitle: String? = null) {
         actionBar.title = title
         actionBar.subtitle = subtitle
